mma_gym/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.17
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
